Revert "Edit"
This reverts commit 7b09f4e7bb20cc331ebe43a9b895bb51f66cf7a1.

+++ audiosr/latent_diffusion/models/ddpm.py
@@ -1564,8 +1564,7 @@ class LatentDiffusion(DDPM):
         energy = torch.cumsum(torch.sum(magnitude, dim=1), dim=1)  # (B, F)
         threshold = energy[:, -1:] * percentile
         # Find first index where energy >= threshold
-        # Find first index where energy >= threshold
-        cutoff_bins = (energy >= threshold).long().argmax(dim=1)
+        cutoff_bins = (energy >= threshold).argmax(dim=1)
         return cutoff_bins
 
     def mel_replace_ops(self, samples, input):
